Added topology extension setup
